+++ src/Category_Colors/Admin.php
@@ -19,7 +19,6 @@ class Admin {
 		add_action( 'admin_notices', array( $this, 'plugin_fail_msg' ) );
 		add_action( 'tribe_settings_below_tabs_tab_category-colors', array( $this, 'is_saved' ) );
 		add_action( 'admin_enqueue_scripts', array( $this, 'load_teccc_js_css' ) );
-		load_plugin_textdomain( 'the-events-calendar-category-colors', false, TECCC_LANG );
 	}
 
 	public function init() {